ROFO-42 createTestImageFile 메소드 위치 이동 -> MultipartFileHelper
테스트 시 여러 기능에서 이미지를 만들 경우가 많아보여
좀 더 공용으로 사용할 수 있는 위치에 뒀습니다.

근데 MultipartFileHelper 라는 이름이 적합한지는 조금 의문이 듭니다.
공용 Fixture 의 이름을 어떻게 지을까 고민하다 정하긴 했지만 더 좋은 이름이 있다면 바꾸는 것도 괜찮아보입니다.
