+++ src/test/kotlin/kr/weit/roadyfoody/support/utils/MultipartFileHelper.kt
@@ -9,3 +9,10 @@ fun createMultipartFile(
     contentStream: InputStream,
     contentType: String = MediaType.APPLICATION_JSON_VALUE,
 ): MockMultipartFile = MockMultipartFile(name, name, contentType, contentStream)
+
+const val TEST_IMAGE_FILE_NAME = "testImage"
+
+fun createTestImageFile(
+    format: ImageFormat,
+    name: String = TEST_IMAGE_FILE_NAME,
+): MockMultipartFile = createMultipartFile(name, generateImageBytes(format).inputStream(), "image/${format.first()}")
